Rename useGetModifiers → createGetModifiers
