+++ src/helpers/createGetModifiers.test.tsx
@@ -1,6 +1,6 @@
-import { DayFlag } from "./UI";
-import { CalendarDay, defaultDateLib } from "./classes/index";
-import { useGetModifiers } from "./useGetModifiers";
+import { DayFlag } from "../UI";
+import { CalendarDay, defaultDateLib } from "../classes/index";
+import { createGetModifiers } from "./createGetModifiers";
 
 const dateLib = defaultDateLib;
 
@@ -40,7 +40,7 @@ const props = {
 describe("useGetModifiers", () => {
   describe("default props", () => {
     // eslint-disable-next-line react-hooks/rules-of-hooks
-    const getModifiers = useGetModifiers(days, props, dateLib);
+    const getModifiers = createGetModifiers(days, props, dateLib);
 
     test("return the modifiers for a given day", () => {
       const modifiers = getModifiers(day2);
@@ -113,7 +113,7 @@ describe("useGetModifiers", () => {
     endMonth.setDate(1);
 
     // eslint-disable-next-line react-hooks/rules-of-hooks
-    const getModifiers = useGetModifiers(
+    const getModifiers = createGetModifiers(
       days,
       { ...props, startMonth, endMonth },
       dateLib
