+++ src/helpers/createGetModifiers.tsx
@@ -1,16 +1,14 @@
-import { DayFlag } from "./UI.js";
-import type { CalendarDay, DateLib } from "./classes/index.js";
-import type { DayPickerProps, Modifiers } from "./types/index.js";
-import { dateMatchModifiers } from "./utils/dateMatchModifiers.js";
+import { DayFlag } from "../UI.js";
+import type { CalendarDay, DateLib } from "../classes/index.js";
+import type { DayPickerProps, Modifiers } from "../types/index.js";
+import { dateMatchModifiers } from "../utils/dateMatchModifiers.js";
 
 /**
  * Return a function to get the modifiers for a given day.
  *
- * NOTE: this is not an hook, but a factory for `getModifiers`.
- *
  * @private
  */
-export function useGetModifiers(
+export function createGetModifiers(
   days: CalendarDay[],
   props: DayPickerProps,
   dateLib: DateLib
